+++ CHANGELOG.md
@@ -5,8 +5,15 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-09-02
+
+### Added
+
+- Show extension snippets in webview.
+- Added a Table of Contents.
+
 ## [0.1.0] - 2020-08-31
 
 ### Added
 
-- Initial release. Shows user and app systems.
+- Initial release. Shows user and app snippets.